Finished the Entity layer

+++ Server/system.DataAccess/Entites/PaymentNoteEntity.cs
@@ -2,14 +2,10 @@
 
 public class PaymentNoteEntity
 {
-    public int Id { get; set; }
-
-    public int BillId { get; set; }
-
+    public long Id { get; set; }
+    public long BillId { get; set; }
     public DateTime Date { get; set; }
-
     public decimal Amount { get; set; }
-
     public string Method { get; set; } = string.Empty;
 
     public BillEntity? Bill { get; set; }
